refactor: 컴포넌트 의존성 명시 및 제공 클래스의 패키지 분리
