+++ di/src/main/java/woowacourse/shopping/di/container/ClassQualifier.kt
@@ -1,0 +1,6 @@
+package woowacourse.shopping.di.container
+
+data class ClassQualifier(
+    val dependencyType: DependencyType,
+    val annotationQualifier: AnnotationQualifier,
+)
